# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CloseConnectionRequest": {
        "description": "Request to close a connection.",
        "properties": {
          "portId": {
            "description": "Target port ID",
            "title": "Portid",
            "type": "string"
          },
          "priority": {
            "default": 1,
            "description": "Command priority",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Priority",
            "type": "integer"
          }
        },
        "required": [
          "portId"
        ],
        "title": "CloseConnectionRequest",
        "type": "object"
      },
...
      "ConnectionInfo": {
        "description": "Connection information.",
        "properties": {
          "baud_rate": {
            "title": "Baud Rate",
            "type": "integer"
          },
          "bytes_read": {
            "title": "Bytes Read",
            "type": "integer"
          },
          "bytes_written": {
            "title": "Bytes Written",
            "type": "integer"
          },
          "connected_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connected At"
          },
          "port_id": {
            "title": "Port Id",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "port_id",
          "status",
          "baud_rate",
          "session_id",
          "bytes_read",
          "bytes_written"
        ],
        "title": "ConnectionInfo",
        "type": "object"
      },
      "ConnectionListResponse": {
        "description": "List of connections.",
        "properties": {
          "connections": {
            "items": {
              "$ref": "#/components/schemas/ConnectionInfo"
            },
            "title": "Connections",
            "type": "array"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "hubId": {
            "title": "Hubid",
            "type": "string"
          }
        },
        "required": [
          "hubId",
          "connections",
          "count"
        ],
        "title": "ConnectionListResponse",
        "type": "object"
      },
...
      "FlashFirmwareRequest": {
        "description": "Request to send flash firmware command.",
        "properties": {
          "artifactFormat": {
            "anyOf": [
              {
                "enum": [
                  "ino",
                  "hex",
                  "bin",
                  "elf"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Firmware format; inferred as ino/hex when omitted (older hubs only accept those)",
            "title": "Artifactformat"
          },
          "boardFqbn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Board FQBN (required for .ino source)",
            "title": "Boardfqbn"
          },
          "boardProfile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Board profile id from the hub's board registry",
            "title": "Boardprofile"
          },
          "firmwareData": {
            "description": "Base64 encoded firmware (.ino source, .hex, .bin or .elf)",
            "title": "Firmwaredata",
            "type": "string"
          },
          "portId": {
            "description": "Target port ID",
            "title": "Portid",
            "type": "string"
          },
          "priority": {
            "default": 3,
            "description": "Command priority",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Priority",
            "type": "integer"
          }
        },
        "required": [
          "portId",
          "firmwareData"
        ],
        "title": "FlashFirmwareRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HubInfo": {
        "description": "Hub information.",
        "properties": {
          "capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "connected": {
            "title": "Connected",
            "type": "boolean"
          },
          "connectedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connectedat"
          },
          "hubId": {
            "title": "Hubid",
            "type": "string"
          },
          "lastSeen": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastseen"
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HubProfile"
              },
              {
                "type": "null"
              }
            ]
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "hubId",
          "connected"
        ],
        "title": "HubInfo",
        "type": "object"
      },
      "HubProfile": {
        "description": "What kind of hub this is, as reported in the handshake.",
        "properties": {
          "mode": {
            "default": "bench",
            "title": "Mode",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "uplink": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uplink"
          }
        },
        "title": "HubProfile",
        "type": "object"
      },
...
      "PortListResponse": {
        "description": "List of ports.",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "hubId": {
            "title": "Hubid",
            "type": "string"
          },
          "ports": {
            "items": {
              "$ref": "#/components/schemas/PortInfo"
            },
            "title": "Ports",
            "type": "array"
          }
        },
        "required": [
          "hubId",
          "ports",
          "count"
        ],
        "title": "PortListResponse",
        "type": "object"
      },
      "RestartDeviceRequest": {
        "description": "Request to send restart device command.",
        "properties": {
          "portId": {
            "description": "Target port ID",
            "title": "Portid",
            "type": "string"
          },
          "priority": {
            "default": 2,
            "description": "Command priority",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Priority",
            "type": "integer"
          }
        },
        "required": [
          "portId"
        ],
        "title": "RestartDeviceRequest",
        "type": "object"
      },
      "SerialWriteRequest": {
        "description": "Request to send serial write command.",
        "properties": {
          "data": {
            "description": "Data to write",
            "title": "Data",
            "type": "string"
          },
          "encoding": {
            "default": "utf-8",
            "description": "Data encoding",
            "title": "Encoding",
            "type": "string"
          },
          "portId": {
            "description": "Target port ID",
            "title": "Portid",
            "type": "string"
          },
          "priority": {
            "default": 5,
            "description": "Command priority",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Priority",
            "type": "integer"
          }
        },
        "required": [
          "portId",
          "data"
        ],
        "title": "SerialWriteRequest",
        "type": "object"
      },
...
      "TaskStatusResponse": {
        "description": "Task status response.",
        "properties": {
          "command_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command Type"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "hub_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hub Id"
          },
          "port_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Port Id"
          },
          "priority": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "progress": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "status",
          "timestamp"
        ],
        "title": "TaskStatusResponse",
        "type": "object"
      },
      "TelemetryEntry": {
        "description": "Telemetry data entry.",
        "properties": {
          "data": {
            "title": "Data",
            "type": "string"
          },
          "dataSizeBytes": {
            "title": "Datasizebytes",
            "type": "integer"
          },
          "portId": {
            "title": "Portid",
            "type": "string"
          },
          "sessionId": {
            "title": "Sessionid",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "timestamp",
          "portId",
          "sessionId",
          "data",
          "dataSizeBytes"
        ],
        "title": "TelemetryEntry",
        "type": "object"
      },
      "TelemetryListResponse": {
        "description": "List of telemetry entries.",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "hubId": {
            "title": "Hubid",
            "type": "string"
          },
          "telemetry": {
            "items": {
              "$ref": "#/components/schemas/TelemetryEntry"
            },
            "title": "Telemetry",
            "type": "array"
          },
          "totalBytes": {
            "title": "Totalbytes",
            "type": "integer"
          }
        },
        "required": [
          "hubId",
          "telemetry",
          "count",
          "totalBytes"
        ],
        "title": "TelemetryListResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Cloud service for testing rpi-hub-service WebSocket communication",
    "title": "RPi Hub Cloud Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/hubs": {
      "get": {
        "description": "List every configured hub (from DEVICE_TOKENS) plus any hub seen since startup,\nwith its connection state.",
        "operationId": "list_hubs_api_hubs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HubInfo"
                  },
                  "title": "Response List Hubs Api Hubs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Hubs",
        "tags": [
          "hubs"
        ]
      }
    },
    "/api/hubs/{hub_id}": {
      "get": {
        "description": "Get hub details.",
        "operationId": "get_hub_api_hubs__hub_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "hub_id",
            "required": true,
            "schema": {
              "title": "Hub Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HubInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Hub",
        "tags": [
          "hubs"
        ]
      }
    },
    "/api/hubs/{hub_id}/commands/close": {
      "post": {
        "description": "Send close connection command to hub.",
        "operationId": "send_close_connection_command_api_hubs__hub_id__commands_close_post",
        "parameters": [
          {
            "in": "path",
            "name": "hub_id",
            "required": true,
            "schema": {
              "title": "Hub Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Send Close Connection Command",
        "tags": [
          "hubs"
        ]
      }
    },
    "/api/hubs/{hub_id}/commands/flash": {
      "post": {
        "description": "Send flash firmware command to hub.",
        "operationId": "send_flash_firmware_command_api_hubs__hub_id__commands_flash_post",
        "parameters": [
          {
            "in": "path",
            "name": "hub_id",
            "required": true,
            "schema": {
              "title": "Hub Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlashFirmwareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Send Flash Firmware Command",
        "tags": [
          "hubs"
        ]
      }
    },
    "/api/hubs/{hub_id}/commands/restart": {
      "post": {
        "description": "Send restart device command to hub.",
        "operationId": "send_restart_device_command_api_hubs__hub_id__commands_restart_post",
        "parameters": [
          {
            "in": "path",
            "name": "hub_id",
            "required": true,
            "schema": {
              "title": "Hub Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestartDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Send Restart Device Command",
        "tags": [
          "hubs"
        ]
      }
    },
    "/api/hubs/{hub_id}/commands/write": {
      "post": {
        "description": "Send serial write command to hub.",
        "operationId": "send_serial_write_command_api_hubs__hub_id__commands_write_post",
        "parameters": [
          {
            "in": "path",
            "name": "hub_id",
            "required": true,
            "schema": {
              "title": "Hub Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SerialWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Send Serial Write Command",
        "tags": [
          "hubs"
        ]
      }
    },
    "/api/hubs/{hub_id}/connections": {
      "get": {
        "description": "Get open serial connections for a hub. Offline hubs report none.",
        "operationId": "get_connections_api_hubs__hub_id__connections_get",
        "parameters": [
          {
            "in": "path",
            "name": "hub_id",
            "required": true,
            "schema": {
              "title": "Hub Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Connections",
        "tags": [
          "hubs"
        ]
      }
    },
    "/api/hubs/{hub_id}/ports": {
      "get": {
        "description": "Get ports for a hub. Offline hubs report no ports.",
        "operationId": "get_ports_api_hubs__hub_id__ports_get",
        "parameters": [
          {
            "in": "path",
            "name": "hub_id",
            "required": true,
            "schema": {
              "title": "Hub Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Ports",
        "tags": [
          "hubs"
        ]
      }
    },
    "/api/hubs/{hub_id}/telemetry": {
      "get": {
        "description": "Get recent telemetry for a hub (kept after the hub disconnects).",
        "operationId": "get_telemetry_api_hubs__hub_id__telemetry_get",
        "parameters": [
          {
            "in": "path",
            "name": "hub_id",
            "required": true,
            "schema": {
              "title": "Hub Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 1000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Telemetry",
        "tags": [
          "hubs"
        ]
      }
    },
...
  },
...
}